Refactor: put EntryOverview's components into own folder
