+++ src/components/entry/overview/icons/ForagingTexas.png
