+++ src/components/entry/overview/icons/Fruitipedia.png
