--- src/components/entry/resources.js
@@ -1,46 +1,0 @@
-import EatTheWeedsLogo from './icons/EatTheWeeds.png'
-import ForagingTexasLogo from './icons/ForagingTexas.png'
-import FruitipediaLogo from './icons/Fruitipedia.png'
-import UrbanMushroomsLogo from './icons/UrbanMushrooms.png'
-import USDALogo from './icons/USDA.svg'
-import WikipediaLogo from './icons/Wikipedia.svg'
-
-/**
- * Resource information for Entry Details. Includes the resource title,
- * url key, logo icon, and icon alt text.
- * @constant {Object[]}
- */
-const RESOURCES = [
-  {
-    title: 'Eat the Weeds',
-    urlKey: 'eat_the_weeds',
-    icon: EatTheWeedsLogo,
-  },
-  {
-    title: 'Foraging Texas',
-    urlKey: 'foraging_texas',
-    icon: ForagingTexasLogo,
-  },
-  {
-    title: 'Fruitipedia',
-    urlKey: 'fruitipedia',
-    icon: FruitipediaLogo,
-  },
-  {
-    title: 'Urban Mushrooms',
-    urlKey: 'urban_mushrooms',
-    icon: UrbanMushroomsLogo,
-  },
-  {
-    title: 'USDA',
-    urlKey: 'usda',
-    icon: USDALogo,
-  },
-  {
-    title: 'Wikipedia',
-    urlKey: 'wikipedia',
-    icon: WikipediaLogo,
-  },
-]
-
-export { RESOURCES }
